requirements-dev.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-braces==1.14.0
django-cors-headers==3.4.0
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.5
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.9
